refactor: 패키지명 변경, ToDo.md 삭제
