+++ src/main/java/lotto/view/InputView.java
@@ -1,4 +1,4 @@
-package lotto.viewer;
+package lotto.view;
 
 import java.util.Arrays;
 import java.util.List;
@@ -7,7 +7,7 @@ import java.util.stream.Collectors;
 
 public class InputView {
 
-    public static final String DELIMITER = ", ";
+    public static final String DELIMITER = ",";
     final Scanner scanner;
 
     public InputView(Scanner scanner) {
@@ -17,11 +17,11 @@ public class InputView {
     public int purchaseMoney() {
         System.out.println("구입금액을 입력해 주세요.");
         return scanner.nextInt();
-    }
 
+    }
     public List<Integer> inputWinningNumbers() {
         System.out.println("지난 주 당첨 번호를 입력해 주세요");
-        scanner.next();
+        scanner.nextLine();
         String rawWinningNumbers = scanner.nextLine();
         String[] splittedWinningNumbers = rawWinningNumbers.split(DELIMITER);
         return Arrays.stream(splittedWinningNumbers)
